+++ web/src/routes/index.ts
@@ -5,7 +5,7 @@ export type RouteValue = (id?: RouteParam) => string
 export type Routes = {
   home: string;
   share: RouteValue;
-  files: RouteValue;
+  category: RouteValue;
   tags: RouteValue;
   works: RouteValue;
   links: RouteValue;
@@ -16,7 +16,7 @@ export type Routes = {
 const routes: Routes = {
   home: '/',
   share: (id?: RouteParam) => id ? `/share/${id}` : '/share',
-  files: (id?: RouteParam) => id ? `/files/${id}` : '/files',
+  category: (id?: RouteParam) => id ? `/category/${id}` : '/category',
   tags: (id?: RouteParam) => id ? `/tags/${id}` : '/tags',
   works: (id?: RouteParam) => id ? `/works/${id}` : '/works',
   links: (id?: RouteParam) => id ? `/links/${id}` : '/links',
